requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==45.0.7
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.5
django-allauth==65.11.2
django-braces==1.17.0
django-celery-beat==2.8.1
django-debug-toolbar==6.0.0
django-embed-video==1.4.10
django-environ==0.12.0
django-extensions @ git+https://github.com/django-extensions/django-extensions.git@25a41d8a3ecb24c009c5f4cac6010a091a3c91c8
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
django-redisboard==9.0.0
django-taggit==6.1.0
django-timezone-field==7.1
html5lib==1.1
idna==3.10
kombu==5.5.4
lxml==6.0.1
MarkupSafe==3.0.2
oauthlib==3.3.1
oscrypto==1.3.0
packaging==25.0
phonenumbers==9.0.13
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pypdf==6.0.0
python-bidi==0.6.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stripe==12.5.1
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.2
xhtml2pdf==0.2.17
